requirements.txt
# =============================================================================
# Project Chimera - Unified Requirements (Legacy Reference)
# For production deployments, use backend-api/requirements.txt with exact pins
# This file uses >= for general compatibility reference only
# MED-005: See backend-api/requirements.txt for canonical dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Security & Authentication
# -----------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0

# Secrets Management (optional - uncomment as needed)
# azure-keyvault-secrets>=4.7.0
# azure-identity>=1.15.0
# boto3>=1.34.0
# hvac>=2.1.0

# -----------------------------------------------------------------------------
# HTTP & Async
# -----------------------------------------------------------------------------
httpx>=0.26.0
aiohttp>=3.9.0
tenacity>=8.2.0

# -----------------------------------------------------------------------------
# AI & LLM Providers
# -----------------------------------------------------------------------------
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.40.0
openai>=1.10.0
anthropic>=0.20.0
# mistralai>=0.1.0

# -----------------------------------------------------------------------------
# NLP & ML
# -----------------------------------------------------------------------------
spacy>=3.7.0
sentence-transformers>=2.2.2
transformers>=4.36.0
accelerate>=0.25.0
faiss-cpu>=1.7.4
datasketch>=1.6.4

# -----------------------------------------------------------------------------
# Red Teaming & LLM Security
# -----------------------------------------------------------------------------
deepteam>=1.0.4
deepeval>=2.5.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.1.0
numpy>=1.26.0

# -----------------------------------------------------------------------------
# Database & Caching
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.0
redis>=5.0.0
aioredis>=2.0.0

# -----------------------------------------------------------------------------
# Monitoring & Observability
# -----------------------------------------------------------------------------
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0

# -----------------------------------------------------------------------------
# Rate Limiting
# -----------------------------------------------------------------------------
slowapi>=0.1.9

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.2.0
httpx>=0.26.0

# Security Testing
bandit>=1.7.0

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
ruff>=0.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8.0
pyyaml>=6.0.0
art>=6.1

# -----------------------------------------------------------------------------
# Production Server
# -----------------------------------------------------------------------------
gunicorn>=21.2.0

# -----------------------------------------------------------------------------
# SpaCy Model (install separately)
# -----------------------------------------------------------------------------
# Run after pip install: python -m spacy download en_core_web_sm
